% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNode.R
\name{addNode}
\alias{addNode}
\title{Add a Distribution}
\usage{
addNode(type, pp, name_short, name_long, Laplace, LaplaceInv, PGF = NULL,
  PGFInv = NULL, simul, cop = NULL, cop_name = "The copula")
}
\arguments{
\item{type}{the type of the node (either 'Mother', 'Child' or 'Both').}

\item{pp}{the parameter of the distribution used in the character strings}

\item{name_short}{the short name of the distribution (ex.: 'log' for the logarithmic distribution)}

\item{name_long}{the long name of the distribution (ex.: 'logarithmic' for the logarithmic distribution)}

\item{Laplace}{the LST of the distribution (character), evaluated at '(z)'}

\item{LaplaceInv}{the inverse LST of the distribution (character), evaluated at '(z)'}

\item{PGF}{the pgf of the distribution (character), if type is 'Mother' or 'Both'}

\item{PGFInv}{the inverse pgf of the distribution (character), if type is 'Mother' or 'Both'}

\item{simul}{function to sample from the distribution}

\item{cop}{function to create a corresponding Archimedean copula class (ex.: for a GEO, it is an AMH copula), can be NULL}

\item{cop_name}{the name used for the corresponding Archimedean copula (generated with cop)}
}
\description{
The function addNode is used to create and add a class of type "Mother" and/or "Child" in the global environment.
}
\examples{
addNode(type = "Child",
        pp = "gamma",
        name_short = "pois",
        name_long = "poisson",
        Laplace = "exp(gamma * (exp(-(z)) - 1))",
        LaplaceInv = "-log(log(z) / gamma + 1)",
        NULL,
        NULL,
        simul = function(n, gamma) rpois(n, gamma),
        cop_name = NULL)

## Construct a bivariate Archimedean copula with the distribution

dist <- POIS(5, 1:2, NULL)
dist@cop(5, 2)

## Add the distribution in a structure

GEO(0.1, NULL, list(GAMMA(0.5, 1:2, NULL),
                    POIS(5, 3:4, NULL)))

}
\author{
Simon-Pierre Gadoury
}

