% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archimedean_Copulas.R
\name{AMH}
\alias{AMH}
\title{Construction of an Archimedean Copula Class Object}
\usage{
AMH(param, dim = 2L)
}
\arguments{
\item{param}{Parameter of the copula}

\item{dim}{Dimension of the copula (>= 2), which is, by default, 2}

\item{density}{Should the expression of the density be computed ?}
}
\description{
Constructs an archm class object with a given parameter and
dimension.
}
\author{
Simon-Pierre Gadoury
}

