% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divers.R
\docType{class}
\name{archm-class}
\alias{archm-class}
\title{#' Shuffle of Min for the Frechet Lower Bound Uniforms
#'
#' Shuffle of min uniforms
#' @param n Number of simulations (numeric)
#' @param a Lower bounds (vector)
#' @param b Upper bounds (vector)
#' @param seed set.seed value
#' @param unif If true, returns the shuffles uniforms as well
#' @importFrom graphics lines
#' @importFrom graphics plot
#' @importFrom graphics points
#' @export}
\keyword{internal}

