% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeCont.R
\name{assuranceVie}
\alias{assuranceVie}
\title{Contrat d'assurance vie discrète}
\usage{
assuranceVie(qx, n = 1, x = 0, m = 0, delta = 0.04, g = 1e+05,
  kappa = 0.99)
}
\arguments{
\item{qx}{qx d'une table de mortalité}
}
\description{
(...)
}

