% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaDeriv.R
\name{BiClayton}
\alias{BiClayton}
\title{Bivariate Clayton Copula}
\usage{
BiClayton(u1, u2, alpha)
}
\arguments{
\item{alpha}{Dependence parameter}
}
\value{
The clayton copula
}
\description{
Bivariate clayton copula function
}

