% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaDeriv.R
\name{BivDeriv}
\alias{BivDeriv}
\title{Bivariate Copula Derivative}
\usage{
BivDeriv(cop)
}
\arguments{
\item{cop}{The copula function}
}
\value{
The symbolical derivative
}
\description{
Take the derivative of bivariate copulas.
}
\details{
Clayton, Frank.
}

