% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaDeriv.R
\name{CompCopBootstrap}
\alias{CompCopBootstrap}
\title{Estimation and bootstrap for one level hierarchical copulas}
\usage{
CompCopBootstrap(m, nn, struc, lower, upper, data = NULL)
}
\arguments{
\item{m}{Number of bootstraps}

\item{nn}{size of the sample per bootstrap}

\item{struc}{Known structure (S4)}

\item{lower}{Lower bound vector for optimization (from left to right, per level)}

\item{upper}{Upper bound vector for optimization (from left to right, per level)}

\item{data}{Data used for the estimation if their is no bootstrap}
}
\description{
Estimation and bootstrap for one level hierarchical copulas
}

