% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaDeriv.R
\name{CompCopEstim}
\alias{CompCopEstim}
\title{Estimation for one level hierarchical copulas}
\usage{
CompCopEstim(data, struc, lower, upper, der)
}
\arguments{
\item{data}{Data used for the estimation}

\item{struc}{Known structure (S4)}

\item{lower}{Lower bound vector for optimization (from left to right, per level)}

\item{upper}{Upper bound vector for optimization (from left to right, per level)}

\item{der}{List of derivatives (from left to right, per level)}
}
\description{
Estimation for one level hierarchical copulas
}

