% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaDeriv.R
\name{CopulaDeriv}
\alias{CopulaDeriv}
\title{Copula derivative}
\usage{
CopulaDeriv(nest, parameters = 1)
}
\arguments{
\item{nest}{The copula function}

\item{parameters}{Number of copulas (if nested)}
}
\value{
The symbolical derivative
}
\description{
Take the derivative of any copula.
}

