% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaDeriv.R
\name{dCompCop}
\alias{dCompCop}
\alias{pCompCop}
\alias{rCompCop}
\alias{rCompCop2}
\title{Density, Cdf, and Random Number Generator for Copulas Constructed Through Compounding}
\usage{
dCompCop(FUN, func = FALSE, code = FALSE, operator = "")

pCompCop(FUN, func = FALSE, code = FALSE, operator = "")

rCompCop(n, FUN, level)

rCompCop2(n, FUN, level)
}
\arguments{
\item{FUN}{Object of class Mother}

\item{func}{If true, returns a function}

\item{code}{If true, copies the LaTeX code to the clipboard}

\item{operator}{Type of cumputer used (only necessary in the case of internal problem)}

\item{n}{Number of realisations}

\item{level}{Number of imbrications}
}
\value{
The copula
}
\description{
Density, Cdf, and Random Number Generator for Copulas Constructed Through Compounding
}
\details{
rCop2 merges rCop and rCompCop2 in one function.
}

