% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaDeriv.R
\name{dCop}
\alias{dCop}
\alias{pCop}
\alias{rCop}
\title{Density, Cdf, and Random Number Generator for Copulas}
\usage{
dCop(copula, vector = TRUE, express = FALSE, code = FALSE,
  operator = "")

pCop(copula, vector = TRUE, express = FALSE, code = FALSE,
  operator = "")

rCop(n, copula)
}
\arguments{
\item{copula}{An Archimedean copula class object}

\item{vector}{If false, returns a function with (x_1, x_2, ..., x_dim, alpha) as arguments.}

\item{express}{If true, returns an expression.}

\item{code}{If true, copies the LaTeX code to clipboard.}

\item{operator}{Type of cumputer used (only necessary in the case of internal problem)}

\item{n}{Number of realisations}

\item{level}{Imbrication levels}

\item{max.dim}{Safe upper bound for final dimension (2000 by default)}
}
\value{
Either an expression, function, code, or sampled data.
}
\description{
Density, Cdf, and Random Number Generator for Copulas
}
\details{
rCop2 combines both the features of rCop and rCompCop in one, allowing the input
of Archimedean copula classes, Mother classes, and Child classes.
}
\author{
Simon-Pierre Gadoury
}

