% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaDeriv.R
\name{density_level1_EXPR}
\alias{density_level1_EXPR}
\title{Expression for the density of a Theta (level 1)}
\usage{
density_level1_EXPR(str, grp, dim, simplify = FALSE)
}
\arguments{
\item{str}{Structure}

\item{grp}{Position of the group}

\item{dim}{Dimension of the group}

\item{simplify}{Should the expression be simply ? (may take time)}
}
\description{
Expression for the density of a Theta (level 1)
}
\details{
The goal is to soon introduce genetic codes to simplify the process
}

