% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaDeriv.R
\name{density_level2_EXPR}
\alias{density_level2_EXPR}
\title{Expression for the density of a Theta (level 2)}
\usage{
density_level2_EXPR(str, grpM, grpB, dim, simplify = FALSE)
}
\arguments{
\item{str}{Structure}

\item{grpM}{Position of the M}

\item{grpB}{Position of the Theta (under M)}

\item{dim}{Dimension of the group}

\item{simplify}{Should the expression be simply ? (may take time)}
}
\description{
Expression for the density of a Theta (level 2)
}
\details{
The goal is to soon introduce genetic codes to simplify the process
}

