% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaDeriv.R
\name{f.mult}
\alias{f.mult}
\title{Bounded Probability for Random Vectors}
\usage{
f.mult(FUN, dim, data, prob = "CDF")
}
\arguments{
\item{FUN}{Multivariate CDF/Survival function}

\item{dim}{Dimension of the random vector}

\item{data}{Matrix of bounds (by row)}

\item{prob}{FUN is a 'CDF' or 'Survival' function ? (either 'CDF' or 'Survival')}
}
\value{
The probability
}
\description{
Porbability computation
}

