% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneticCodes.R
\name{GeneticCodes}
\alias{GeneticCodes}
\title{Obtain the genetic codes of a structure}
\usage{
GeneticCodes(str)
}
\arguments{
\item{str}{The structure}
}
\value{
A list of of the structure's genetic codes.
}
\description{
Obtain the genetic codes of a structure
}
\examples{
str <- GEO(0.5, NULL, list(GAMMA(1/30, c(5,6), NULL),
                           GEO(0.1, NULL, list(GAMMA(1/30, c(1,2), NULL),
                                               GAMMA(1/30, c(3,4), NULL)))))
GeneticCodes(str)

}
\author{
Simon-Pierre Gadoury
}

