% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archimedean_Copulas.R
\name{Gumbel}
\alias{Gumbel}
\title{Construction of an Archimedean Copula Class Object}
\usage{
Gumbel(param, dim = 2L)
}
\arguments{
\item{param}{Parameter of the copula}

\item{dim}{Dimension of the copula (>= 2), which is, by default, 2}

\item{density}{Should the expression of the density be computed ?}
}
\value{
An archm S4 class object
}
\description{
Constructs a Gumbel Archimedean copula object with a given parameter and dimension
}
\author{
Simon-Pierre Gadoury
}

