% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvLap_Child.R
\name{InvLap_Child}
\alias{InvLap_Child}
\title{Inverse LST of a Child Node}
\usage{
InvLap_Child(code, str, tt = "z")
}
\arguments{
\item{code}{Genetic code of the child node (can be a leaf i.e. end by 0)}

\item{str}{Object of class Mother (the structure)}

\item{tt}{Output variable to be used ('z' by default)}
}
\description{
Inverse LST of a Child Node
}
\author{
Simon-Pierre Gadoury
}

