% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divers.R
\name{maxvCopula}
\alias{maxvCopula}
\title{MLE for copulas}
\usage{
maxvCopula(start, data, low, up, FUN, gradient = NULL, param.type = "alpha")
}
\arguments{
\item{start}{Initial parameter}

\item{data}{Data used for the estimation}

\item{low}{Lower bound for the research}

\item{up}{Upper bound fo the research}

\item{FUN}{The density of the copula to estimate (an expression)}

\item{gradient}{(Not yet included)}

\item{param.type}{The parameters to estimate (character)}
}
\value{
The estimation.
}
\description{
Estimate the parameters of a copula.
}
\seealso{
\code{\link{var.attrib}}
}

