% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_Code.R
\name{Moth}
\alias{Moth}
\alias{Moth2}
\alias{Chil}
\title{Show Method Function for Compounding}
\usage{
Moth(object, indent = "", delta.indent = 3, label = NA)

Moth2(object, indent = "", delta.indent = 3, label = NA)

Chil(object, space = "")
}
\arguments{
\item{object}{S4 object}

\item{indent}{Initial spacing for children}

\item{label}{Internal use (must not be changed)}

\item{delta.ident}{Added space for each iteration}
}
\description{
Show Method Function for Compounding
}
\keyword{internal}
