% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divers.R
\name{pairs2}
\alias{pairs2}
\title{Extension of pairs}
\usage{
pairs2(x, cex = 1, labels = paste("u", 1:length(x[1, ]), sep = ""),
  cex.labels = 1, digits = 2)
}
\arguments{
\item{x}{Matrix of data}

\item{cex}{Scale of the graph}

\item{labels}{Vector of character (names of the variables)}

\item{cex.labels}{Numeric. The size of the labels' font}

\item{digits}{Number of digits to be shown for Kendall's tau}
}
\value{
The pair graphs with respective Kendall tau
}
\description{
Matrix of scatterplots with respective Kendall's tau
}
\seealso{
\code{\link{pairs}}
}

