% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCompCop2.R
\name{pCompCop2}
\alias{pCompCop2}
\title{Construction of the copula with a known structure}
\usage{
pCompCop2(str)
}
\arguments{
\item{str}{Object of class Mother}
}
\value{
An expression in terms of u
}
\description{
Construction of the copula with a known structure
}
\details{
This is a test.
}
\author{
Simon-Pierre Gadoury
}

