% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCop.R
\name{pCop}
\alias{pCop}
\title{Cdf, and Random Number Generator for Copulas}
\usage{
pCop(copula, vector = TRUE)
}
\arguments{
\item{copula}{An Archimedean copula class object}

\item{vector}{If false, returns a function with (u_1, u_2, ..., u_dim) as arguments, else, just (u)}
}
\value{
Either an expression, function, code, or sampled data.
}
\description{
Cdf, and Random Number Generator for Copulas
}
\author{
Simon-Pierre Gadoury
}

