% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaDeriv.R
\name{PGF.Power}
\alias{PGF.Power}
\title{Derivative of the power of a pgf}
\usage{
PGF.Power(tt, n, s, str)
}
\arguments{
\item{tt}{What should be the expression evaluated to ? (the variable)}

\item{n}{The number of derivatives}

\item{s}{The power}

\item{str}{The structure}
}
\description{
Derivative of the power of a pgf
}

