% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCompCop.R
\name{rCompCop}
\alias{rCompCop}
\title{Density, Cdf, and Random Number Generator for Copulas Constructed Through Compounding}
\usage{
rCompCop(n, FUN, level)
}
\arguments{
\item{n}{Number of realisations}

\item{FUN}{Object of class Mother}

\item{level}{Number of imbrications}
}
\description{
The density, the CDF, and the random number generator for copulas construted
through compounding.
}
\details{
rCompCop2 is more general (and easier to use) than rCompCop, but is slower.
}
\author{
Simon-Pierre Gadoury
}

