% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCop.R
\name{rCop}
\alias{rCop}
\title{Density, Cdf, and Random Number Generator for Copulas}
\usage{
rCop(n, copula)
}
\arguments{
\item{n}{Number of realisations}

\item{copula}{An Archimedean copula class object}
}
\value{
Sampled data.
}
\description{
Density, Cdf, and Random Number Generator for Copulas
}
\author{
Simon-Pierre Gadoury
}

