% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeCont.R
\name{rentes}
\alias{rentes}
\title{Contrat de rentes discrètes}
\usage{
rentes(qx, n = 1, x = 0, m = 0, delta = 0.04, g = 10000,
  kappa = 0.99)
}
\arguments{
\item{qx}{qx d'une table de mortalité}
}
\description{
(...)
}

