% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApproxSum.R
\name{repBiv}
\alias{repBiv}
\title{Bivariate Dependent Sum Approximation (CDF)}
\usage{
repBiv(s, m, FUN, dep)
}
\arguments{
\item{s}{CDF evaluation}

\item{m}{Precision}

\item{FUN}{The copula}

\item{dep}{Dependence parameter}
}
\value{
The lower and upper bound approximation of the CDF at s
}
\description{
Bivariate lower and upper bound approximation
}

