% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rStruc.R
\name{rStruc}
\alias{rStruc}
\title{Sampling for compound rvs}
\usage{
rStruc(n, str)
}
\arguments{
\item{n}{Number of realisations}

\item{str}{Object of class Mother (the structure)}
}
\description{
Sampling for compound rvs
}
\author{
Simon-Pierre Gadoury
}

