% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_Code.R
\docType{methods}
\name{show,archm-method}
\alias{show,archm-method}
\title{Show Method for Copulas}
\usage{
\S4method{show}{archm}(object)
}
\description{
Show Method for Copulas
}
\keyword{Show}
\keyword{internal}
