% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_Code.R
\docType{methods}
\name{show,Child-method}
\alias{show,Child-method}
\title{Show Method for Compounding Functions - Child Class}
\usage{
\S4method{show}{Child}(object)
}
\description{
Show Method for Compounding Functions - Child Class
}
\keyword{Show}
\keyword{internal}
