% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_Code.R
\docType{methods}
\name{show,Mother-method}
\alias{show,Mother-method}
\title{Show Method for Compounding Functions - Mother Class}
\usage{
\S4method{show}{Mother}(object)
}
\description{
Show Method for Compounding Functions - Mother Class
}
\keyword{Show}
\keyword{internal}
