% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divers.R
\name{shuff}
\alias{shuff}
\title{Shuffle of Min for the Frechet Lower Bound Uniforms}
\usage{
shuff(n, a, b, seed = sample(1:2e+05, 1), unif = FALSE)
}
\arguments{
\item{n}{Number of simulations (numeric)}

\item{a}{Lower bounds (vector)}

\item{b}{Upper bounds (vector)}

\item{seed}{set.seed value}

\item{unif}{If true, returns the shuffles uniforms as well}
}
\description{
Shuffle of min uniforms
}

