% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeCont.R
\docType{data}
\name{t2052}
\alias{t2052}
\title{SOA canadian mortality tables}
\format{An object of class \code{data.frame} with 107 rows and 2 columns.}
\usage{
data("t2052")
}
\description{
Table 2052.
}
\keyword{datasets}

