% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeCont.R
\docType{data}
\name{t2061}
\alias{t2061}
\title{SOA canadian mortality tables}
\format{An object of class \code{data.frame} with 111 rows and 2 columns.}
\usage{
data("t2061")
}
\description{
Table 2061.
}
\keyword{datasets}

