% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeCont.R
\docType{data}
\name{t2790}
\alias{t2790}
\title{SOA canadian mortality tables}
\format{An object of class \code{data.frame} with 98 rows and 2 columns.}
\usage{
data("t2790")
}
\description{
Table 2790.
}
\keyword{datasets}

