% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divers.R
\name{var.attrib}
\alias{var.attrib}
\title{Automatic assigner}
\usage{
var.attrib(type, data, output = "x", input = "k", use = "param")
}
\arguments{
\item{type}{Type of attribution: 0 = vector, 1 = row and 2 = column}

\item{data}{Data used (has to be the same as input)}

\item{output}{Desired variable for attribution}

\item{input}{Data variable: vector (0) or matrix (1 or 2)}
}
\value{
Vector of attribution (expression)
}
\description{
Assign vector of values to variables automatically.
}
\keyword{internal}

