% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMH_Copulas.R
\name{AMH}
\alias{AMH}
\title{Construction of an Archimedean Copula Class Object.}
\usage{
AMH(param, dim = 2L, density = FALSE)
}
\arguments{
\item{param}{parameter of the copula.}

\item{dim}{dimension of the copula (>= 2), which is, by default, 2.}

\item{density}{compute the expression of the density of the copulas.}
}
\value{
An archm S4 class object.
}
\description{
Constructs an AMH Archimedean copula object with
a given parameter and dimension.
}
\details{
Constructs an AMH Archimedean copula object with a given parameter and dimension.
}
\author{
Simon-Pierre Gadoury
}
