% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clayton_Copulas.R
\name{Clayton}
\alias{Clayton}
\title{Construction of an Archimedean Copula Class Object}
\usage{
Clayton(param, dim = 2L, density = FALSE)
}
\arguments{
\item{param}{the parameter of the copula.}

\item{dim}{the dimension of the copula (>= 2), which is, by default, 2.}

\item{density}{logical. Should the expression of the density of the copula be computed?}
}
\value{
An archm S4 class object.
}
\description{
Constructs a Clayton Archimedean copula object with
a given parameter and dimension.
}
\author{
Simon-Pierre Gadoury
}
