% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaDeriv.R
\name{Clayton}
\alias{AMH}
\alias{Clayton}
\alias{Frank}
\alias{Gumbel}
\title{Construction of an Archimedean Copula Class Object}
\usage{
Clayton(param, dim = 2L, density = FALSE)

Frank(param, dim = 2L)

AMH(param, dim = 2L)

Gumbel(param, dim = 2L)
}
\arguments{
\item{param}{Parameter of the copula}

\item{dim}{Dimension of the copula (>= 2), which is, by default, 2}
}
\description{
Constructs an archm class object with a given parameter and
dimension.
}

