% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Structure_Distributions.R
\name{GAMMA}
\alias{GAMMA}
\title{Construction of a Child Class Object}
\usage{
GAMMA(par, unif, struc = NULL)
}
\arguments{
\item{par}{Dimension of the distribution}

\item{unif}{Uniform structure, a numeric vector of grouped
numbers.

i.e. c(1,2,3) is translated as being c(u1, u2, u3).}

\item{struc}{Nesting structure of the form

X(par1, c(i,...), list(Y(par2, c(j,...), NULL),
                       Z(par3, c(k,...), NULL))),

where X, Y, and Z are compatible functions (see 'details').
It is to note that if struc is NULL, the function will automatically
be of class Child. For continuous distributions (i.e. GAMMA), struc is
always NULL.}
}
\description{
Constructs a Child class object for
a given parameter and arguments
}
\section{Slots}{

\describe{
\item{\code{Param}}{The name of the parameter used}

\item{\code{parameter}}{The value of the parameter}

\item{\code{dimension}}{The dimension}

\item{\code{type}}{The type of function (either child or mother)}

\item{\code{arguments}}{The corresponding arguments (ex.: arguments 1 and 2 imply 'u1' and 'u2')}

\item{\code{structure}}{The structure below the node of type 'Mother'}

\item{\code{Laplace}}{Expression of the LST}

\item{\code{LaplaceInv}}{Expression of the inverse LST}

\item{\code{PGF}}{Expression of the pgf}

\item{\code{PGFInv}}{Expression of the inverse pgf}

\item{\code{simul}}{Fonction to sample from the distribution}

\item{\code{theta}}{I don't know honestly}

\item{\code{cop}}{Construct an Archimedean copula with this distribution}

\item{\code{Der}}{Fonction to compute the expression of the 'k'th derivative of either the 'PGF', 'PGFInv', 'Laplace' or 'LaplaceInv'}

\item{\code{FUN}}{Fonction to compute the function of the 'k'th derivative of either the 'PGF', 'PGFInv', 'Laplace' or 'LaplaceInv'}
}}
\examples{
GEO(0.5, NULL, list(GAMMA(1/30, c(5,6), NULL),
                    GEO(0.1, NULL, list(GAMMA(1/30, c(1,2), NULL),
                                        GAMMA(1/30, c(3,4), NULL)))))
}
\author{
Simon-Pierre Gadoury
}

