% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaDeriv.R
\name{GeneticCodes}
\alias{GeneticCodes}
\title{Obtain the genetic codes of a structure}
\usage{
GeneticCodes(str)
}
\arguments{
\item{str}{The structure}
}
\description{
Obtain the genetic codes of a structure
}

