% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaDeriv.R
\name{GeneticCodes}
\alias{GeneticCodes}
\title{#' Derivative of the power of a pgf
#'
#' @param tt What should be the expression evaluated to ? (the variable)
#' @param n The number of derivatives
#' @param s The power
#' @param str The structure
#'
#' @export}
\usage{
GeneticCodes(str)
}

