% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneticCodes.R
\name{GeneticCodes}
\alias{GeneticCodes}
\title{Obtain the genetic codes of a structure}
\usage{
GeneticCodes(str)
}
\arguments{
\item{str}{The structure}
}
\value{
A list of of the structure's genetic codes.
}
\description{
Obtain the genetic codes of a structure
}
\author{
Simon-Pierre Gadoury
}

