% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneticCodes.R
\name{GeneticCodes}
\alias{GeneticCodes}
\title{Obtain the Genetic Codes of a Structure}
\usage{
GeneticCodes(str)
}
\arguments{
\item{str}{an object of class Mother (the structure)}
}
\value{
A list of the structure's genetic codes.
}
\description{
Function to obtain the list of all genetic codes of a structure.
}
\examples{
## Create the structure
str <- GEO(0.5, NULL, list(GAMMA(1/30, c(5,6), NULL),
                           GEO(0.1, NULL, list(GAMMA(1/30, c(1,2), NULL),
                                               GAMMA(1/30, c(3,4), NULL)))))
## Get the genetic codes
GeneticCodes(str)

}
\author{
Simon-Pierre Gadoury
}
