% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gumbel_Copulas.R
\name{Gumbel}
\alias{Gumbel}
\title{Construction of an Archimedean Copula Class Object}
\usage{
Gumbel(param, dim = 2L)
}
\arguments{
\item{param}{parameter of the copula}

\item{dim}{dimension of the copula (>= 2), which is, by default, 2}
}
\value{
An archm S4 class object.
}
\description{
Constructs a Gumbel Archimedean copula object with
a given parameter and dimension.
}
\author{
Simon-Pierre Gadoury
}
