% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lap_Child.R
\name{Lap_Child}
\alias{Lap_Child}
\title{LST of a Child Node}
\usage{
Lap_Child(code, str, tt = "z")
}
\arguments{
\item{code}{Genetic code of the child node (can be a leaf i.e. end by 0)}

\item{str}{Object of class Mother (the structure)}

\item{tt}{Output variable to be used ('z' by default)}
}
\value{
...
}
\description{
The function Lap_Child() with a specific path and a predefined structure
(S4 calss of a type 'Mother'), returns the Laplace-Stieltjes transform expression of
the corresponding node with a specific variable.
}
\examples{

str <- GEO(0.1, NULL, list(GAMMA(0.1, 1:2, NULL),
                           GAMMA(0.2, 3:4, NULL)))

InvLap_Child(c(0,2), str)

}
\author{
Simon-Pierre Gadoury
}
\seealso{
\link{InvLap_Child}
}

