% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lap_Child.R
\name{Lap_Child}
\alias{Lap_Child}
\title{LST of a Child Node}
\usage{
Lap_Child(code, str, tt = "z")
}
\arguments{
\item{code}{Genetic code of the child node (can be a leaf i.e. end by 0)}

\item{str}{Object of class Mother (the structure)}

\item{tt}{Output variable to be used ('z' by default)}
}
\description{
With a specific path and a predefined structure (S4 class of a type 'Mother'), this function returns the LST
expression of the corresponding node with a specific variable.
}
\author{
Simon-Pierre Gadoury
}
\seealso{
\link{InvLap_Child}
}

