% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Node.R
\name{Node}
\alias{Node}
\title{Obtain a node with its genetic code}
\usage{
Node(path, str)
}
\arguments{
\item{path}{Genetic code of the node}

\item{str}{The structure}
}
\value{
The node distribution, the child dimension, the parameter and the composition.
}
\description{
Obtain a node with its genetic code.
}
\examples{
# We directly give the path of the desired node.
Node(c(0,2,2), LOG(0.5, NULL, list(GAMMA(1/30, c(5,6), NULL),
                              LOG(0.1, NULL, list(GAMMA(1/30, c(1,2), NULL),
                              GAMMA(1/30, c(3,4), NULL))))))

# Here we provide the path with the GeneticCodes function of this package.
Node(GeneticCodes(LOG(0.5, NULL, list(GAMMA(1/30, c(5,6), NULL),
                                     LOG(0.1, NULL, list(GAMMA(1/30, c(1,2), NULL),
                                     GAMMA(1/30, c(3,4), NULL))))))[[3]],
                                 LOG(0.5, NULL, list(GAMMA(1/30, c(5,6), NULL),
                                 LOG(0.1, NULL, list(GAMMA(1/30, c(1,2), NULL),
                                 GAMMA(1/30, c(3,4), NULL))))))

}
\author{
Simon-Pierre Gadoury
}

