% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Node.R
\name{Node}
\alias{Node}
\title{Obtain a node with its genetic code}
\usage{
Node(path, str)
}
\arguments{
\item{path}{Genetic code of the node}

\item{str}{The structure}
}
\description{
Obtain a node with its genetic code
}
\author{
Simon-Pierre Gadoury
}

