% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCompCop.R
\name{pCompCop}
\alias{pCompCop}
\title{Construction of the Copula With a Known Structure}
\usage{
pCompCop(str)
}
\arguments{
\item{str}{Object of class Mother}
}
\value{
An expression in terms of u.
}
\description{
The function pCompCop() construct a copula with a know structure.
}
\examples{
pCompCop(LOG(0.5, NULL, list(GAMMA(1/30, c(5,6), NULL),
                              LOG(0.1, NULL, list(GAMMA(1/30, c(1,2), NULL),
                              GAMMA(1/30, c(3,4), NULL))))))
}
\author{
Simon-Pierre Gadoury
}

