% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCompCop.R
\name{pCompCop}
\alias{pCompCop}
\title{Construction of the copula with a known structure}
\usage{
pCompCop(str)
}
\arguments{
\item{str}{Object of class Mother}
}
\value{
An expression in terms of u.
}
\description{
The function pCompCop() construct a copula with a know structure.
}
\author{
Simon-Pierre Gadoury
}

