% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCop.R
\name{pCop}
\alias{pCop}
\title{Cdf, and Random Number Generator for Copulas}
\usage{
pCop(copula, vector = TRUE, express = FALSE, code = FALSE,
  operator = "")
}
\arguments{
\item{copula}{An Archimedean copula class object}

\item{vector}{If false, returns a function with (x_1, x_2, ..., x_dim, alpha) as arguments.}

\item{express}{If true, returns an expression.}

\item{code}{If true, copies the LaTeX code to clipboard.}

\item{operator}{Type of cumputer used (only necessary in the case of internal problem)}
}
\value{
Either an expression, function, code, or sampled data.
}
\description{
Cdf, and Random Number Generator for Copulas
}
\author{
Simon-Pierre Gadoury
}

