% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCop.R
\name{pCop}
\alias{pCop}
\title{Cdf function for Archimedean copulas (archm) class objects}
\usage{
pCop(copula, vector = TRUE, express = FALSE)
}
\arguments{
\item{copula}{An Archimedean copula (archm) class object}

\item{vector}{If false, returns a function with (u_1, u_2, ..., u_dim) as arguments, else,
just (u)}
}
\value{
A function ...
}
\description{
The cdf function for Archimedean copulas (arcm) class objetcs.
}
\examples{
cop <- Clayton(5, 2)
pCop(cop, vector = TRUE)(c(0.1, 1))
pCop(cop, vector = FALSE)(0.1, 0.1)

}
\author{
Simon-Pierre Gadoury
}
\seealso{
\link{rCop}, \link{Clayton}, \link{AMH}, \link{Gumbel}, \link{Frank}
}

