% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCompCop.R
\name{rCompCop}
\alias{rCompCop}
\title{Random number generator for Mother class objects}
\usage{
rCompCop(n, str)
}
\arguments{
\item{n}{the number of realisations.}

\item{str}{an object of class Mother.}
}
\value{
A numeric matrix of sampled data from the structure
}
\description{
Samples from a Mother class object.
}
\examples{
## Create the structure
str <- GEO(0.1, 1, list(GAMMA(0.2, 2:3, NULL),
                        GEO(0.3, 4:5, NULL)))

## Sample from the structure
rCompCop(1000, str)

}
\author{
Simon-Pierre Gadoury
}
