% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCompCop2.R
\name{rCompCop2}
\alias{rCompCop2}
\title{Density, Cdf, and Random Number Generator for Copulas Constructed Through Compounding}
\usage{
rCompCop2(n, str)
}
\arguments{
\item{n}{Number of realisations}

\item{str}{Object of class Mother}
}
\description{
Density, Cdf, and Random Number Generator for Copulas Constructed Through Compounding
}
\details{
rCompCop2 is more general (and easier to use) than rCompCop, but is slower.
}
\author{
Simon-Pierre Gadoury
}

