% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rStruc.R
\name{rStruc}
\alias{rStruc}
\title{Sampling for compound rvs}
\usage{
rStruc(n, str)
}
\arguments{
\item{n}{Number of realisations}

\item{str}{Object of class Mother (the structure)}
}
\value{
Return...
}
\description{
The function rStruc() generate n samples from the compound rvs.
}
\examples{
rStruc(10000, GEO(0.5, NULL, list(GAMMA(1/30, c(5,6), NULL),
                                     GEO(0.1, NULL, list(GAMMA(1/30, c(1,2), NULL),
                                     GAMMA(1/30, c(3,4), NULL))))))

}
\author{
Simon-Pierre Gadoury
}

