% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rStruc.R
\name{rStruc}
\alias{rStruc}
\title{Sampling From Compound Random Variables}
\usage{
rStruc(n, str)
}
\arguments{
\item{n}{the number of realisations.}

\item{str}{a S4 object of class Mother (the structure).}
}
\value{
A numeric matrix of sampled values from the specified structure.
For every child node, the sampled values are repeated d times, where d
is the length of the node's 'unif' argument.
}
\description{
Generate n samples from a structure of compound rvs.
}
\examples{
rStruc(10000, GEO(0.5, NULL, list(GAMMA(1/30, c(1), NULL),
                                     GEO(0.1, NULL, list(GAMMA(1/30, c(2), NULL),
                                     GAMMA(1/30, c(3), NULL))))))

}
\author{
Simon-Pierre Gadoury
}
